Added all tests for the data class in a single test file
